Split EmbeddedContentHelper into a thin facade and a new DataSource.
